Tidy up command folder
