+++ src/BlazorDatasheet.Core/Commands/Data/MergeCellsCommand.cs
@@ -1,9 +1,8 @@
 using BlazorDatasheet.Core.Data;
 using BlazorDatasheet.Core.Data.Cells;
 using BlazorDatasheet.DataStructures.Geometry;
-using BlazorDatasheet.Formula.Core;
 
-namespace BlazorDatasheet.Core.Commands;
+namespace BlazorDatasheet.Core.Commands.Data;
 
 public class MergeCellsCommand : IUndoableCommand
 {
