+++ src/BlazorDatasheet.Core/Commands/Data/SortRangeCommand.cs
@@ -4,7 +4,7 @@ using BlazorDatasheet.DataStructures.Geometry;
 using BlazorDatasheet.DataStructures.Store;
 using BlazorDatasheet.Formula.Core;
 
-namespace BlazorDatasheet.Core.Commands;
+namespace BlazorDatasheet.Core.Commands.Data;
 
 public class SortRangeCommand : IUndoableCommand
 {
