+++ src/BlazorDatasheet.Core/Commands/Formatting/SetFormatCommand.cs
@@ -1,12 +1,10 @@
 using BlazorDatasheet.Core.Data;
 using BlazorDatasheet.Core.Data.Cells;
-using BlazorDatasheet.Core.Events;
 using BlazorDatasheet.Core.Formats;
 using BlazorDatasheet.DataStructures.Geometry;
 using BlazorDatasheet.DataStructures.Intervals;
-using BlazorDatasheet.DataStructures.Store;
 
-namespace BlazorDatasheet.Core.Commands;
+namespace BlazorDatasheet.Core.Commands.Formatting;
 
 public class SetFormatCommand : IUndoableCommand
 {
