+++ src/BlazorDatasheet.Core/Commands/RowCols/UnhideRowsCommand.cs
@@ -1,16 +1,16 @@
-using BlazorDatasheet.Core.Data;
+﻿using BlazorDatasheet.Core.Data;
 using BlazorDatasheet.DataStructures.Geometry;
 
-namespace BlazorDatasheet.Core.Commands;
+namespace BlazorDatasheet.Core.Commands.RowCols;
 
-public class HideRowsCommand : IUndoableCommand
+public class UnhideRowsCommand : IUndoableCommand
 {
     private int _start;
     private int _end;
-
+    
     private RowInfoStoreRestoreData _restoreData = null!;
 
-    public HideRowsCommand(int start, int end)
+    public UnhideRowsCommand(int start, int end)
     {
         _start = start;
         _end = end;
@@ -18,14 +18,14 @@ public class HideRowsCommand : IUndoableCommand
 
     public bool Execute(Sheet sheet)
     {
-        _restoreData = sheet.Rows.HideRowsImpl(_start, _end - _start + 1);
+        _restoreData = sheet.Rows.UnhideRowsImpl(_start, _end - _start + 1);
         return true;
     }
 
     public bool Undo(Sheet sheet)
     {
         sheet.Rows.Restore(_restoreData);
-        sheet.MarkDirty(new RowRegion(_start, _end));
+        sheet.MarkDirty(new RowRegion(_start,_end));
         return true;
     }
 }
