+++ src/BlazorDatasheet.Core/Data/Cells/CellStore.Data.cs
@@ -1,4 +1,5 @@
 using BlazorDatasheet.Core.Commands;
+using BlazorDatasheet.Core.Commands.Data;
 using BlazorDatasheet.DataStructures.Store;
 using BlazorDatasheet.Formula.Core;
