+++ src/BlazorDatasheet.Core/Data/RowInfoStore.cs
@@ -1,4 +1,5 @@
 using BlazorDatasheet.Core.Commands;
+using BlazorDatasheet.Core.Commands.RowCols;
 using BlazorDatasheet.Core.Data.Cells;
 using BlazorDatasheet.Core.Events.Layout;
 using BlazorDatasheet.Core.Formats;
