+++ src/BlazorDatasheet.Core/Events/BeforeRangeSortEventArgs.cs
@@ -1,5 +1,7 @@
 ﻿using System.ComponentModel;
 using BlazorDatasheet.Core.Commands;
+using BlazorDatasheet.Core.Commands.Data;
+using BlazorDatasheet.Core.Commands.RowCols;
 using BlazorDatasheet.DataStructures.Geometry;
 
 namespace BlazorDatasheet.Core.Events;
