+++ src/BlazorDatasheet.Core/Validation/ValidationManager.cs
@@ -1,4 +1,5 @@
 ﻿using BlazorDatasheet.Core.Commands;
+using BlazorDatasheet.Core.Commands.Formatting;
 using BlazorDatasheet.Core.Data;
 using BlazorDatasheet.Core.Events.Validation;
 using BlazorDatasheet.Core.Interfaces;
