+++ src/BlazorDatasheet.SharedPages/Pages/Index.razor
@@ -10,6 +10,8 @@
 @using BlazorDatasheet.Core.Validation
 @using System.Drawing
 @using BlazorDatasheet.Core.Commands
+@using BlazorDatasheet.Core.Commands.Data
+@using BlazorDatasheet.Core.Commands.RowCols
 
 <PageTitle>Index</PageTitle>
