+++ src/BlazorDatasheet/SelectionMenu.razor
@@ -1,4 +1,6 @@
 @using BlazorDatasheet.Core.Commands
+@using BlazorDatasheet.Core.Commands.Data
+@using BlazorDatasheet.Core.Commands.RowCols
 @using BlazorDatasheet.Core.Data
 @using BlazorDatasheet.DataStructures.Geometry
 @using BlazorDatasheet.Menu
