+++ test/BlazorDatasheet.Test/Commands/AutofillCommandTests.cs
@@ -1,5 +1,6 @@
 using System.Linq;
 using BlazorDatasheet.Core.Commands;
+using BlazorDatasheet.Core.Commands.Data;
 using BlazorDatasheet.Core.Data;
 using BlazorDatasheet.Core.Formats;
 using BlazorDatasheet.Core.Interfaces;
