+++ test/BlazorDatasheet.Test/Commands/ChangeCellValueCommandTests.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using BlazorDatasheet.Core.Commands;
+using BlazorDatasheet.Core.Commands.Data;
 using BlazorDatasheet.Core.Data;
 using NUnit.Framework;
