+++ test/BlazorDatasheet.Test/Commands/ClearCellsCommandTests.cs
@@ -1,4 +1,5 @@
 ﻿using BlazorDatasheet.Core.Commands;
+using BlazorDatasheet.Core.Commands.Data;
 using BlazorDatasheet.Core.Data;
 using BlazorDatasheet.DataStructures.Geometry;
 using BlazorDatasheet.Formula.Core;
