+++ test/BlazorDatasheet.Test/Commands/HideRowsCommandTests.cs
@@ -1,4 +1,5 @@
 using BlazorDatasheet.Core.Commands;
+using BlazorDatasheet.Core.Commands.RowCols;
 using BlazorDatasheet.Core.Data;
 using FluentAssertions;
 using NUnit.Framework;
