+++ test/BlazorDatasheet.Test/Commands/MergeCellsAndInsertColRowTests.cs
@@ -1,4 +1,5 @@
 ﻿using BlazorDatasheet.Core.Commands;
+using BlazorDatasheet.Core.Commands.RowCols;
 using BlazorDatasheet.Core.Data;
 using BlazorDatasheet.DataStructures.Geometry;
 using NUnit.Framework;
