+++ test/BlazorDatasheet.Test/Formula/SheetFormulaIntegrationTests.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using BlazorDatasheet.Core.Commands;
+using BlazorDatasheet.Core.Commands.Data;
 using BlazorDatasheet.Core.Data;
 using BlazorDatasheet.Core.FormulaEngine;
 using BlazorDatasheet.DataStructures.Geometry;
