+++ test/BlazorDatasheet.Test/SheetTests/FormattingTests.cs
@@ -1,5 +1,6 @@
 using System.Linq;
 using BlazorDatasheet.Core.Commands;
+using BlazorDatasheet.Core.Commands.RowCols;
 using BlazorDatasheet.Core.Data;
 using BlazorDatasheet.Core.Formats;
 using BlazorDatasheet.DataStructures.Geometry;
